made changes to Car file inside components folder renamed some functions commented out code inside Computers and any reference to it
