+++ src/components/Hero.jsx
@@ -2,7 +2,8 @@
 import { motion } from "framer-motion";
 
 import { styles } from "../styles";
-import { ComputersCanvas, CarCanvas } from "./canvas";
+// import { ComputersCanvas } from "./canvas";
+import { CarCanvas } from "./canvas";
 
 const Hero = () => {
   return (
@@ -28,7 +29,7 @@ const Hero = () => {
         </div> */}
       </div>
 
-      <ComputersCanvas />
+      <CarCanvas />
 
       <div className='absolute xs:bottom-10 bottom-32 w-full flex justify-center items-center'>
         <a href='#about'>
