+++ src/components/canvas/index.js
@@ -1,7 +1,8 @@
 import EarthCanvas from "./Earth";
 import BallCanvas from "./Ball";
-import ComputersCanvas from "./Computers";
+// import ComputersCanvas from "./Computers";
 import StarsCanvas from "./Stars";
 import CarCanvas from "./Car";
 
-export { EarthCanvas, BallCanvas, ComputersCanvas, StarsCanvas, CarCanvas };
+export { EarthCanvas, BallCanvas, StarsCanvas, CarCanvas };
+// export { EarthCanvas, BallCanvas, ComputersCanvas, StarsCanvas };
